task(SDK-5584) - Refactors Media Handling for InApps (#955)
* task(SDK-5584) - Refactors Media Handling for InApps

* task(SDK-5584) - Reverts hardcoded type

* task(SDK-5584) - Removes hideImageViewForNonImageMedia config

* task(SDK-5584) - Removes fillVideoFrame config

* task(SDK-5584) - Reverts order in cleanup for footerfragment

* task(SDK-5584) - Removes redundant onPause from GIFHandler

* task(SDK-5584) - Adds default lifecycle observer to prevent multiple forwards each time

* task(SDK-5584) - Removes InAppMediaDelegate.kt
