+++ clevertap-core/src/main/java/com/clevertap/android/sdk/inapp/fragment/CTInAppNativeCoverImageFragment.kt
@@ -7,26 +7,27 @@ import android.view.ViewGroup
 import android.widget.FrameLayout
 import android.widget.RelativeLayout
 import androidx.core.graphics.toColorInt
-import androidx.media3.common.util.UnstableApi
 import com.clevertap.android.sdk.R
+import com.clevertap.android.sdk.inapp.media.InAppMediaConfig
+import com.clevertap.android.sdk.inapp.media.InAppMediaHandler
 import com.clevertap.android.sdk.applyInsetsWithMarginAdjustment
 import com.clevertap.android.sdk.customviews.CloseImageView
 
-@UnstableApi
 internal class CTInAppNativeCoverImageFragment : CTInAppBaseFullFragment() {
 
-    private lateinit var mediaDelegate: InAppMediaDelegate
+    private lateinit var mediaHandler: InAppMediaHandler
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        mediaDelegate = InAppMediaDelegate(
+        mediaHandler = InAppMediaHandler.create(
             fragment = this,
             inAppNotification = inAppNotification,
             currentOrientation = currentOrientation,
             isTablet = inAppNotification.isTablet && isTablet(),
-            resourceProvider = resourceProvider()
+            resourceProvider = resourceProvider(),
+            supportsStreamMedia = true
         )
-        mediaDelegate.initVideoPlayerHandle()
+        lifecycle.addObserver(mediaHandler)
     }
 
     override fun onCreateView(
@@ -47,10 +48,9 @@ internal class CTInAppNativeCoverImageFragment : CTInAppBaseFullFragment() {
 
         val relativeLayout = fl.findViewById<RelativeLayout>(R.id.cover_image_relative_layout)
 
-        mediaDelegate.bindVideoFrame(relativeLayout.findViewById(R.id.video_frame))
-        mediaDelegate.setMediaForInApp(
+        mediaHandler.setup(
             relativeLayout,
-            InAppMediaConfig(imageViewId = R.id.cover_image, clickableMedia = true),
+            InAppMediaConfig(imageViewId = R.id.cover_image, clickableMedia = true, videoFrameId = R.id.video_frame),
             CTInAppNativeButtonClickListener()
         )
 
@@ -58,7 +58,7 @@ internal class CTInAppNativeCoverImageFragment : CTInAppBaseFullFragment() {
 
         closeImageView.setOnClickListener {
             didDismiss(null)
-            mediaDelegate.clearGif()
+            mediaHandler.clear()
             activity?.finish()
         }
 
@@ -71,28 +71,9 @@ internal class CTInAppNativeCoverImageFragment : CTInAppBaseFullFragment() {
         return inAppView
     }
 
-    override fun onStart() {
-        super.onStart()
-        mediaDelegate.onStart()
-    }
-
-    override fun onResume() {
-        super.onResume()
-        mediaDelegate.onResume()
-    }
-
-    override fun onPause() {
-        super.onPause()
-        mediaDelegate.onPause()
-    }
-
-    override fun onStop() {
-        super.onStop()
-        mediaDelegate.onStop()
-    }
-
     override fun cleanup() {
+        lifecycle.removeObserver(mediaHandler)
+        mediaHandler.cleanup()
         super.cleanup()
-        mediaDelegate.cleanup()
     }
 }
